requirements.txt
asgiref==3.6.0
Django==4.2
sqlparse==0.4.4
psycopg2-binary==2.9.1
python-dotenv==0.18.0
Cerberus==1.3.3
PyJWT==1.7.1
sentry-sdk==1.9.10